+++ _sample_suites/simple/test_1.py
@@ -14,6 +14,18 @@ def test_2():
 def test_3():
     slash.skip_test('skipped')
 
+
+@slash.parametrize('param1', ['string', {'some': 'dict'}])
+def test_parametrized(param1, some_parameter):  # pylint: disable=unused-argument
+    pass
+
+
+@slash.fixture
+@slash.parametrize('param', ['very long' * 1000, 2, 'third arg'])
+def some_parameter(param):
+    return param
+
+
 class SampleClassTest(slash.Test):
 
     def test_method(self):
@@ -21,7 +33,7 @@ class SampleClassTest(slash.Test):
 
 
 @slash.parametrize('param', ['x.y', 'x(y'])
-def test_4(param): # pylint: disable=unused-argument
+def test_4(param):  # pylint: disable=unused-argument
     pass